requirements.txt
matplotlib==3.4.2
tqdm==4.61.1
pandas
gdown
torch==1.8.1+cu111 
torchvision==0.9.1+cu111
pillow==8.2.0